+++ Sources/VideoKit/Pagination/Paginable.swift
@@ -1,0 +1,12 @@
+//
+//  Paginable.swift
+//  VideoKit
+//
+//  Created by Daniel Saidi on 2025-11-18.
+//  Copyright © 2025 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/// This protocol can be implemented by paginable item types.
+public protocol Paginable: Hashable, Identifiable, Sendable {}